fix: use just one method to retrieve mnemonic and infura key external variables
