+++ solidity/truffle-examples/random-datasource/test/utils.js
@@ -11,15 +11,16 @@ const waitForEvent = (_event, _from = 0, _to = 'latest') =>
 
 const fileExists = _path => fs.existsSync(_path)
 
-const getInfuraKey = () => fileExists(API_PATH)
-  ? require(API_PATH)['infuraKey']
-  : process.env['infuraKey']
-  ? process.env['infuraKey']
-  : (console.log(`Cannot test! Please provide 'infuraKey' as an environment variable, or export it from '${API_PATH}'!`), process.exit(1))
+const getExternalVariable = _variable =>
+  fileExists(API_PATH)
+    ? require(API_PATH)[_variable]
+    : process.env[_variable]
+    ? process.env[_variable]
+    : (console.log(`Cannot test! Please provide '${_variable}' as an environment variable, or export it from '${API_PATH}'!`), process.exit(1))
 
 module.exports = {
   waitForEvent,
   PREFIX,
-  getInfuraKey,
+  getExternalVariable,
   fileExists
 }
